+++ src/App.tsx
@@ -2,15 +2,11 @@ import React, { useState, useEffect, useMemo } from "react";
 import { themes, type ThemeName } from "./styles/theme";
 import type { Movie } from "./types/Movie";
 import Header from "./components/Header";
+import MovieForm from "./components/MovieForm";
 
 function App() {
   const [themeName, setThemeName] = useState<ThemeName>("light");
   const [movies, setMovies] = useState<Movie[]>([]);
-  const [newTitle, setNewTitle] = useState<string>("");
-  const [newDirector, setNewDirector] = useState<string>("");
-  const [newYear, setNewYear] = useState<number | "">("");
-  const [newGenre, setNewGenre] = useState<string>("");
-  const [newRating, setNewRating] = useState<number | "">("");
   const [searchTerm, setSearchTerm] = useState<string>("");
   const [error, setError] = useState<string | null>(null);
   const [isLoading, setIsLoading] = useState<boolean>(false);
@@ -44,38 +40,19 @@ function App() {
     );
   }, [movies, searchTerm]);
 
-  const handleAddMovie = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!newTitle || !newDirector || !newYear || !newGenre || !newRating) {
-      setError("모든 필드를 입력해주세요.");
-      return;
-    }
-
-    const newMovie = {
-      title: newTitle,
-      director: newDirector,
-      year: Number(newYear),
-      genre: newGenre,
-      rating: Number(newRating),
-    };
-
+  const handleAddMovie = async (movie: Omit<Movie, "id">) => {
     try {
       setIsLoading(true);
       const res = await fetch("/api/movies", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(newMovie),
+        body: JSON.stringify(movie),
       });
 
       if (!res.ok) throw new Error("영화를 추가하지 못했습니다.");
 
       const savedMovie: Movie = await res.json();
       setMovies((prev) => [...prev, savedMovie]);
-      setNewTitle("");
-      setNewDirector("");
-      setNewYear("");
-      setNewGenre("");
-      setNewRating("");
       setError(null);
     } catch (err: unknown) {
       if (err instanceof Error) {
@@ -112,105 +89,7 @@ function App() {
           {error}
         </div>
       )}
-      <div
-        style={{
-          marginBottom: "24px",
-          padding: "20px",
-          borderRadius: "12px",
-          backgroundColor: currentTheme.componentBg,
-          border: `1px solid ${currentTheme.border}`,
-        }}
-      >
-        <h2>영화 추가</h2>
-        <form
-          onSubmit={handleAddMovie}
-          style={{ display: "flex", gap: "12px", flexWrap: "wrap" }}
-        >
-          <input
-            type="text"
-            placeholder="제목"
-            value={newTitle}
-            onChange={(e) => setNewTitle(e.target.value)}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-            }}
-          />
-          <input
-            type="text"
-            placeholder="감독"
-            value={newDirector}
-            onChange={(e) => setNewDirector(e.target.value)}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-            }}
-          />
-          <input
-            type="number"
-            placeholder="연도"
-            value={newYear}
-            onChange={(e) => setNewYear(Number(e.target.value))}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-              width: "80px",
-            }}
-          />
-          <input
-            type="text"
-            placeholder="장르"
-            value={newGenre}
-            onChange={(e) => setNewGenre(e.target.value)}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-            }}
-          />
-          <input
-            type="number"
-            placeholder="평점"
-            value={newRating}
-            onChange={(e) => {
-              const val = Number(e.target.value);
-              if (val >= 0 && val <= 10) setNewRating(val);
-            }}
-            style={{
-              padding: "8px",
-              borderRadius: "8px",
-              border: `1px solid ${currentTheme.border}`,
-              backgroundColor: currentTheme.inputBg,
-              color: currentTheme.text,
-              width: "100px",
-            }}
-          />
-          <button
-            type="submit"
-            style={{
-              padding: "8px 16px",
-              backgroundColor: currentTheme.buttonBg,
-              color: currentTheme.buttonText,
-              border: "none",
-              borderRadius: "8px",
-              cursor: "pointer",
-            }}
-          >
-            추가
-          </button>
-        </form>
-      </div>
+      <MovieForm themeName={themeName} onAddMovie={handleAddMovie} />
       <div
         style={{
           padding: "20px",
